+++ scripts/macros/2014/classFeatures/paladin/auraOfProtection.js
@@ -3,7 +3,8 @@ async function create({trigger: {entity: item, target, identifier}}) {
     if (itemUtils.getConfig(item, 'combatOnly') && !combatUtils.inCombat()) return;
     let targetEffect = effectUtils.getEffectByIdentifier(target.actor, identifier);
     if (targetEffect) {
-        if (targetEffect.origin === item.uuid) return;
+        let effectOriginItem = await effectUtils.getOriginItem(targetEffect);
+        if (effectOriginItem.uuid === item.uuid) return;
         await genericUtils.remove(targetEffect);
     }
     let showIcon = itemUtils.getConfig(item, 'showIcon');